Moved dashboard button configuration to DashboardRepository
