--- public/dashboard.php.rej
@@ -1,34 +1,0 @@
-diff a/public/dashboard.php b/public/dashboard.php	(rejected hunks)
-@@ -13,19 +13,19 @@ $allButtons = $controller->getButtons();
- <img class="banner-img" src="images/bannerImg.jpg" alt="Banner afbeelding" />
- 
- <section id="dashboard-container" class="flex justify-center">
--    <div id="dashboardBox" class="col-6">    
--        <div class="employeeBox flex justify-center p-30">
--            <?php if (isset($user) && ($user->getRole() === UserRole::EMPLOYEE || $user->getRole() === UserRole::ADMINISTRATOR)) {
--                    // Er wordt over elke button geloopt en deze wordt op de pagina weergegeven
--                    echo '<p class="dasboardSectionTitel col-12 pb-30">EMPLOYEE</p>';
--                    foreach($allButtons as $button) {
--                        if ($button['role'] === "EMPLOYEE") {
--                            echo '<a class="dashboardButton col-3 text-center" href="'.$button['url'].'">' .$button['nameButton']. '</a>';
--                        }
--                    }
--                }
--        ?>
--        </div>
-+    <div id="dashboardBox" class="col-6">
-+		<?php if (isset($user) && ($user->getRole() === UserRole::EMPLOYEE || $user->getRole() === UserRole::ADMINISTRATOR)): ?>
-+			<div class="employeeBox flex justify-center p-30">
-+			<!-- Er wordt over elke button geloopt en deze wordt op de pagina weergegeven -->
-+				<p class="dasboardSectionTitel col-12 pb-30">EMPLOYEE</p>
-+				<?php foreach($allButtons as $button) {
-+					if ($button['role'] === "EMPLOYEE") {
-+						echo '<a class="dashboardButton col-3 text-center" href="'.$button['url'].'">' .$button['nameButton']. '</a>';
-+					}
-+				}?>
-+			</div>
-+        <?php endif; ?>
-+
-         <hr class="lineBetweenRoles col-12"></hr> 
-         <div class="administratorBox flex justify-center p-30">
-             <!-- Met alleen de rol ADMINISTRATOR wordt de sectie administrator getoond -->
